+++ src/main/java/com/mealtiger/backend/rest/controller/RecipeController.java
@@ -4,7 +4,8 @@ import com.mealtiger.backend.database.model.recipe.Rating;
 import com.mealtiger.backend.database.model.recipe.Recipe;
 import com.mealtiger.backend.rest.error_handling.exceptions.EntityNotFoundException;
 import com.mealtiger.backend.rest.error_handling.exceptions.RatingOwnRecipeException;
-import com.mealtiger.backend.rest.model.recipe.RatingResponse;
+import com.mealtiger.backend.rest.model.rating.AverageRatingResponse;
+import com.mealtiger.backend.rest.model.rating.RatingRequest;
 import com.mealtiger.backend.rest.model.recipe.RecipeRequest;
 import com.mealtiger.backend.database.repository.RecipeRepository;
 import com.mealtiger.backend.rest.model.recipe.RecipeResponse;
